% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidABM.R
\docType{data}
\name{age_brackets}
\alias{age_brackets}
\title{Age brackets}
\format{
A data frame with columns:
\describe{
\item{lower}{Lower bound for bracket}
\item{upper}{Upper bound for bracket}
\item{bracket}{Description of the bracket}
}
}
\usage{
age_brackets
}
\description{
Age brackets
}
\keyword{internal}
