% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{build_agent_df}
\alias{build_agent_df}
\title{Create one or more agents}
\usage{
build_agent_df(
  age,
  sex,
  med_cond,
  seir_status,
  transmission_params,
  progression_params,
  sympt = NULL,
  ticks = NULL
)
}
\arguments{
\item{age}{Age in years.}

\item{sex}{Sex: "M" or "F" or a factor with levels "M" and "F".}

\item{med_cond}{Logical: has a pre-existing medical condition}

\item{seir_status}{SEIR status: "S", "E", "I", "R", or an ordered factor
with those levels.}

\item{transmission_params}{A data frame with transmission parameters for
agents}

\item{progression_params}{A data frame with parameters for the timing of
disease progression (from "E" to "I" and "I" to "R")}

\item{sympt}{Subject will be symptomatic if they are infected (logical).}

\item{ticks}{For exposed or infected agents: Number of ticks since they
entered that state.}
}
\value{
A data.table of agent characteristics with columns
\itemize{
\item \code{age}: Age in years (numeric).
\item \code{age_bkt}: The age bracket (ordered factor).
\item \code{sex}: The sex (factor with levels "M" and "F").
\item \code{med_cond}: Has a medical condition (logical).
\item \code{sympt}: Has symptoms of COVID-19, if infectious (logical).
\item \code{seir}: SEIR status (ordered factor with levels "S", "E", "I", and "R").
\item \code{id}: Unique ID number (integer)
\item \code{ticks}: days since reaching current disease stage (relevant to "E" and
"I" status only) (integer).
\item \code{target}: target ticks for transition to next compartment (relevant to
"E" and "I" status only).
\item \code{x_shed}: Shedding parameter (relevant to "I" status only)
\item \code{x_susc}: Susceptibility parameter (relevant to "S" status only)
\item \code{shape_ei}: Shape parameter for progression from "E" to "I" compartment.
\item \code{scale_ei}: Scale parameter for progression from "E" to "I" compartment.
\item \code{shape_ir}: Shape parameter for progression from "I" to "R" compartment.
\item \code{scale_ir}: Scale parameter for progression from "I" to "R" compartment.
}
}
\description{
Create agents and initialize with properties.
}
\examples{
# ADD_EXAMPLES_HERE
}
