% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{build_prob_matrix}
\alias{build_prob_matrix}
\title{Build a probability matrix for transmission of infection}
\usage{
build_prob_matrix(prob_df)
}
\arguments{
\item{prob_df}{A data frame with columns: Subject refers to the susceptible
subject and agent refers to the infectious agent.
\itemize{
\item \code{sub_age_bkt} An ordered factor for age-bracket (lower = younger)
\item \code{sub_sex} A factor or character vector "F" or "M"
\item \code{sub_med_cond} Logical: Has pre-existing condition
\item \code{agt_age_bkt} An ordered factor for age-bracket.
\item \code{agt_sex}  A factor or character vector "F" or "M"
\item \code{agt_sympt} Logical: Does the agent display symptoms?
\item \code{prob} The probability of transmitting the infection from agent to
subject.
}}
}
\value{
A 6-dimensional matrix.
}
\description{
Take a data frame with columns corresponding to discrete coordinates in a
multidimensional space and form a matrix.
}
\examples{
# ADD_EXAMPLES_HERE
}
