% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{build_transition_matrix}
\alias{build_transition_matrix}
\title{Build a probability matrix for transitions between compartnemts}
\usage{
build_transition_matrix(trans_df)
}
\arguments{
\item{prob_df}{A data frame with columns:
\itemize{
\item \code{age_bkt} An ordered factor for age-bracket (lower = younger)
\item \code{sex} A factor or character vector "F" or "M"
\item \code{med_cond} Logical: Has pre-existing condition
\item \code{sympt} Logical: Does the agent display symptoms if infected?
\item \code{prob} The probability of transitioning to the next compartment
}}
}
\value{
A 4-dimensional matrix.
}
\description{
Take a data frame with columns corresponding to discrete coordinates in a
multidimensional space and form a matrix.
}
\examples{
# ADD_EXAMPLES_HERE
}
