% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{create_agents}
\alias{create_agents}
\title{Create one or more agents}
\usage{
create_agents(age, sex, med_cond, seir_status, sympt, fix_sympt = TRUE)
}
\arguments{
\item{age}{Age in years.}

\item{sex}{Sex: "M" or "F" or a factor with levels "M" and "F".}

\item{med_cond}{Logical: has a pre-existing medical condition}

\item{seir_status}{SEIR status: "S", "E", "I", "R", or an ordered factor
with those levels.}

\item{sympt}{Subject is infected and has symptoms (logical).}

\item{fix_sympt}{Set \code{sympt} to \code{FALSE} for agents where \code{seir_status} is
not "I".}
}
\value{
A tibble of agent characteristics with columns
\itemize{
\item \code{age}: Age in years (numeric).
\item \code{age_bkt}: The age bracket (ordered factor).
\item \code{sex}: The sex (factor with levels "M" and "F").
\item \code{med_cond}: Has a medical condition (logical).
\item \code{sympt}: Has symptoms of COVID-19, if infectious (logical).
\item \code{seir}: SEIR status (ordered factor with levels "S", "E", "I", and "R").
\item \code{id}: Unique ID number (integer)
\item \code{ticks}: days since reaching current disease stage (relevant to "E" and
"I" status) (integer).
}
}
\description{
Create agents and initialize with properties.
}
\examples{
# ADD_EXAMPLES_HERE
}
