% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{decode_seir}
\alias{decode_seir}
\title{Translate integer index to SEIR level}
\usage{
decode_seir(x)
}
\arguments{
\item{x}{An integer (or vector) with values 1--4.}
}
\value{
A factor with levels "S", "E", "I", or "R"
}
\description{
Translate an integer index to an SEIR factor
}
\examples{
decode_seir(3)

}
