% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agents.R
\name{generate_agent_params}
\alias{generate_agent_params}
\title{Build a list of parameters for constructing agents}
\usage{
generate_agent_params(
  agent_count,
  age_dist,
  p_female,
  seir_dist,
  p_med_cond,
  p_sympt
)
}
\arguments{
\item{agent_count}{The number of agents to generate parameters for.}

\item{age_dist}{A function to sample from a probability distribution or
else a vector or list of ages. The vector will be repeated if
necessary to produce enough agents.}

\item{p_female}{A probability of being female or else a vector or list of
sexes ("M" or "F") or logical (\code{TRUE} for female, \code{FALSE} for male).
The vector will be repeated if necessary to produce enough agents.}

\item{seir_dist}{A function to sample from a probability distribution or
else a vector or list of SEIR statuses ("S", "E", "I", or "R"). The vector
will be repeated if necessary to produce enough agents.}

\item{p_med_cond}{A probability of having a medical condition or else a
vector or list of medical conditions (logical: \code{TRUE} or \code{FALSE}).
The vector will be repeated if necessary to produce enough agents.}

\item{p_sympt}{A probability of being symptomatic if infectious, or else
a vector or list of symptomatic status if infectious
(logical: \code{TRUE} or \code{FALSE}). The vector will be repeated if necessary to
produce enough agents.}
}
\value{
A named list of parameters for \code{age}, \code{sex}, \code{seir}, \code{med_cond},
and \code{sympt}.
}
\description{
Generate parameters for constructing agents. Given distributions for
ages, sexes, medical conditions, SEIR status, and symptomatic conditions
for infectious individuals, generate a list of parameters for constructing
a set of agents.
}
\examples{
# ADD_EXAMPLES_HERE
}
