% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_age_bracket}
\alias{get_age_bracket}
\title{Get the age bracket}
\usage{
get_age_bracket(x)
}
\arguments{
\item{x}{Age in years (vector)}
}
\value{
Ordered factor indicating the age bracket
}
\description{
Convert an age (in years) to a bracket.
}
\examples{
get_age_bracket(21)
get_age_bracket(c(10, 20, 21, 25, 29, 30, 31, 39, 40, 41))
}
