% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{get_prob}
\alias{get_prob}
\title{Look up a disease-transmission probability}
\usage{
get_prob(agent, neighbors, probs)
}
\arguments{
\item{agent}{An agent (a row of an agents data frame or a vertex in an
agent network) with \code{seir_status} = "I".}

\item{neighbors}{A list of neighbors (a data frame or list of neighboring
vertices from the network graph).}

\item{probs}{A probability matrix generated by \code{build_prob_matrix}.}
}
\value{
A vector of probabilities corresponding to the list of neighbors.
}
\description{
Look up a disease-transmission probability from an infectious agent to a
susceptible subject, based on the characteristics of agent and subject.
}
\examples{
# ADD_EXAMPLES_HERE
}
