% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_seir_level}
\alias{get_seir_level}
\title{Translate a letter into an integer SEIR level}
\usage{
get_seir_level(x)
}
\arguments{
\item{x}{a character or character vector with values "S", "E", "I", or "R".}
}
\value{
An integer value
}
\description{
Translate a character SEIR status into an integer index equivalent to the
factor level
}
\examples{
get_seir_level("I")
}
