% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{get_transition}
\alias{get_transition}
\title{Look up a transition probability to the next compartment}
\usage{
get_transition(agent, transitions)
}
\arguments{
\item{agent}{An agent (a row of an agents data frame or a vertex in an
agent network) with \code{seir_status} = "I".}

\item{neighbors}{A list of neighbors (a data frame or list of neighboring
vertices from the network graph).}

\item{probs}{A transition probability matrix generated by
\code{build_transition_matrix}.}
}
\value{
A vector of transition probabilities.
}
\description{
Look up a probability for an agent in compartment E or I to transition to
the next compartment (I or R, respectively), based on the characteristics of
agent and subject.
}
\examples{
# ADD_EXAMPLES_HERE
}
