% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect.R
\name{infect}
\alias{infect}
\title{Propagate infections on the network}
\usage{
infect(agents, neighbors, x0 = -2.556)
}
\arguments{
\item{agents}{A \code{data.table} containing agent information.}

\item{neighbors}{A list of \code{data.table}s, each containing the neighbors on
a different network.
Neighbor data.tables have columns:
\itemize{
\item \code{head}: the head vertex
\item \code{tail}: the tail vertex
\item \code{edge_type}: the kind of network (e.g., "home", "social", or "work")
\item \code{freq}: the frequency of contacts on this edge (times per day the two
agents are in close contact).
\item \code{strength}: the intensity/closeness of the contact (e.g., health-care
visits are very strong, paying a cashier at a store would be rather
weak).
}}

\item{x0}{The unmodified mean infection value..}
}
\value{
A modified network graph after propagating the infections.
}
\description{
For each infected node on the network, find susceptible neighbors and
probabilistically spread the infection.
}
\examples{
# ADD_EXAMPLES_HERE
}
