% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidABM.R
\docType{data}
\name{prog_df}
\alias{prog_df}
\title{Progression parameters for disease progression}
\format{
A named list of five-dimensional arrays of doubles
}
\usage{
prog_df
}
\description{
Progression parameters for disease progression
}
\keyword{internal}
