% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_prog.R
\name{progress_disease}
\alias{progress_disease}
\title{Progress agents in the "E" and "I" compartments to the next compartments
("I" and "R", respectively).}
\usage{
progress_disease(agents)
}
\arguments{
\item{agents}{A \code{data.table} containing agents. This \code{data.table} should have
columns:
\itemize{
\item \code{seir}: The SEIR status (integer index ranging from 1 to 4)
\item \code{ticks}: The number of time steps the agent has been in the compartment.
\item \code{ei_shape}, \code{ei_scale}, \code{ir_shape}, \code{ir_scale}: shape and scale
parameters for the probability distriubtions for transitions from the
current compartment to the next compartment for "E" to "I" and "I" to
"R", respectively.
}}
}
\value{
Invisibly returns a modified version of the \code{agents} data table.
}
\description{
Progress agents in the "E" and "I" compartments to the next compartments
("I" and "R", respectively).
}
\examples{
# ADD_EXAMPLES_HERE
}
