% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_probs}
\alias{read_probs}
\title{Read probabilities for disease transmission}
\usage{
read_probs(prob_file = "probabilities.csv")
}
\arguments{
\item{prob_file}{A \code{.csv} file with the transmission probabilities.}
}
\value{
An n-dimensional matrix of probabilities
}
\description{
Read disease transmission probabilities from a file and transform to
a probability matrix.
}
\examples{
# ADD_EXAMPLES_HERE
}
