% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_trans}
\alias{read_trans}
\title{Read parameters for disease progression}
\usage{
read_trans(trans_file = "transitions.csv")
}
\arguments{
\item{trans_file}{A \code{.csv} file with the transition probabilities.}
}
\value{
A named list containing two n-dimensional matrices of parameters
(with elements \code{e} for exposed to infectious transitions and \code{i} for
infectious to recovered transitions).
}
\description{
Read parameters for disease progression and create transition probability
matrices for transitions from exposed to infectious and from infectious to
recovered. The parameters characterize the mean and scale of the
distribution (in days).
}
\examples{
# ADD_EXAMPLES_HERE
}
