% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run a model}
\usage{
run_model(model, ticks)
}
\arguments{
\item{model}{A model object returned from \link{create_model}.}

\item{ticks}{Number of days to simulate.}
}
\value{
A named list with the resulting model after the ticks, and
a tibble with the counts of agents in "S", "E", "I", and "R" status.
}
\description{
Run a model for a specified number of ticks, where each tick is a day.
}
\examples{
# ADD_EXAMPLES_HERE
}
