% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{setup_model}
\alias{setup_model}
\title{Set up a model}
\usage{
setup_model(
  agent_count,
  age_dist,
  p_female,
  seir_dist,
  p_med_cond,
  p_symptomatic,
  trans_df = NULL,
  prog_df = NULL,
  home_nw_params = NULL,
  social_nw_params = NULL,
  work_nw_params = NULL
)
}
\arguments{
\item{agent_count}{The number of agents to generate parameters for.}

\item{age_dist}{A function to sample from a probability distribution or
else a vector or list of ages. The vector will be repeated if
necessary to produce enough agents.}

\item{p_female}{A probability of being female or else a vector or list of
sexes ("M" or "F") or logical (\code{TRUE} for female, \code{FALSE} for male).
The vector will be repeated if necessary to produce enough agents.}

\item{seir_dist}{A function to sample from a probability distribution or
else a vector or list of SEIR statuses ("S", "E", "I", or "R"). The vector
will be repeated if necessary to produce enough agents.}

\item{p_med_cond}{A probability of having a medical condition or else a
vector or list of medical conditions (logical: \code{TRUE} or \code{FALSE}).
The vector will be repeated if necessary to produce enough agents.}

\item{p_symptomatic}{A probability of being symptomatic if infectious, or
else a vector or list of symptomatic status if infectious
(logical: \code{TRUE} or \code{FALSE}). The vector will be repeated if necessary to
produce enough agents.}

\item{trans_df}{A data frame of parameters for disease
transmission probability distributions. Columns are:
\itemize{
\item \code{age_bkt}: The age bracket (ordered factor).
\item \code{sex}: The sex (factor with levels "M" and "F").
\item \code{med_cond}: Has a medical condition (logical).
\item \code{sympt}: Has symptoms of COVID-19, if infectious (logical).
\item \code{mu_shed}: Mean of the shedding parameter distribution.
\item \code{sigma_shed}: Scale (standard deviation) of the shedding parameter
distribution.
\item \code{mu_susc}: Mean of the susceptibility parameter distribution
\item \code{sigma_susc}: Scale (standard deviation) of the susceptibility parameter
distribution.
}}

\item{prog_df}{A data frame of parameters for disease
progression probability distributions. Columns are:
\itemize{
\item \code{age_bkt}: The age bracket (ordered factor).
\item \code{sex}: The sex (factor with levels "M" and "F").
\item \code{med_cond}: Has a medical condition (logical).
\item \code{sympt}: Has symptoms of COVID-19, if infectious (logical).
\item \code{shape_ei}: Shape parameter for progression from "E" to "I" compartment.
\item \code{scale_ei}: Scale parameter for progression from "E" to "I" compartment.
\item \code{shape_ir}: Shape parameter for progression from "I" to "R" compartment.
\item \code{scale_ir}: Scale parameter for progression from "I" to "R" compartment.
}}

\item{home_nw_params}{A named list with parameters for the home-contacts
network. The list should have elements:
\itemize{
\item \code{freq}: The frequency of contacts in the network (contacts per day)
\item \code{intens}: The intensity of contacts in the network
\item \code{topol}: The topology of the network: "smallworld" or "Barabasi Albert".
\item \code{nbrw}: The size of the neighborhood (small world) or the number of edges
to add per time step (BA).
\item \code{p_rewire}: The probability of rewiring an edge (small world) or the
power of the preferential attachments (BA).
}}

\item{social_nw_params}{A named list with parameters for social contacts.}

\item{work_nw_params}{A named list with parameters for work contacts.}
}
\value{
A named list of the model: agents, network, contagion
probabilities for disease transmission, and transition probabilities for
progress to the next stage of the disease.
}
\description{
Initialize a model (agents and network).
}
\examples{
# ADD_EXAMPLES_HERE
}
