% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agents.R
\name{setup_test_model}
\alias{setup_test_model}
\title{FUNCTION_TITLE}
\usage{
setup_test_model(
  n_agents = 100,
  mean_age = 40,
  std_age = 20,
  p_female = 0.5,
  p_med_cond = 0.2,
  p_seir = c(0.99, 0, 0.01, 0),
  p_sympt = 0.5,
  neighbors = 5,
  p_rewire = 0.05
)
}
\arguments{
\item{n_agents}{The number of agents in the model.}

\item{mean_age}{The mean age of an agent.}

\item{std_age}{The standard deviation of ages among agents.}

\item{p_female}{The fraction of agents who are female.}

\item{p_med_cond}{Probability of having a medical condition.}

\item{p_seir}{A vector of probabilities for "S", "E", "I", and "R" status.
These should add up to 1.0.}

\item{p_sympt}{Probability of being symptomatic when infectious.}

\item{neighbors}{The number of neighbor edges in the ring network graph to
initialize the small world network (2 = connect to both nearest neighbors
on the ring;  4 = connect to both neighbors and their neighbors, etc.).}

\item{p_rewire}{The probability to rewire an edge from a neighbor to a
random node..}
}
\value{
A model, as returned by \link{setup_model}
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
