% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick.R
\name{tick}
\alias{tick}
\title{Advance model by one tick}
\usage{
tick(model, probs, transitions)
}
\arguments{
\item{model}{A model object.}

\item{probs}{A probability array (n-dimensional) for transmission of
infection from infectious to susceptible agents.}

\item{An}{array of transition probabilities between compartments of the
disease.}
}
\value{
An updated model
}
\description{
Advance model by one tick (i.e., one day). Calculate progress of disease
(exposed people become infectious, and infectious people recover) and then
calculate transmission of the disease (stochastically, infectious people may
infect susceptible people) through network connections.
}
\examples{
# ADD_EXAMPLES_HERE
}
