% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidABM.R
\docType{data}
\name{trans}
\alias{trans}
\title{Transition parameters for disease progression}
\format{
A named list of five-dimensional arrays of doubles
}
\usage{
trans
}
\description{
Transition parameters for disease progression
}
\keyword{internal}
