% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidABM.R
\docType{data}
\name{trans_df}
\alias{trans_df}
\title{Disease transmission probability arraw}
\format{
A six-dimensional array of doubles
}
\usage{
trans_df
}
\description{
Disease transmission probability arraw
}
\keyword{internal}
