% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_prog.R
\name{transition_prob}
\alias{transition_prob}
\title{Look up a progression probability to the next compartment}
\usage{
transition_prob(ticks, shape, scale)
}
\arguments{
\item{ticks}{The number of time steps in which the agent has been in the
current compartment}

\item{shape}{The shape parameter for the transition probability distribution}

\item{scale}{The scale parameter for the transition probability distribution}
}
\value{
A vector of progression probabilities.
}
\description{
Look up a probability for an agent in compartment E or I to progress to
the next compartment (I or R, respectively), based on the characteristics of
agent and subject.
}
\examples{
# ADD_EXAMPLES_HERE
}
