% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{infect}
\alias{infect}
\title{Propagate infections on the network}
\usage{
infect(network, prob, tracing = FALSE)
}
\arguments{
\item{network}{An igraph network, generated by \code{create_network}}

\item{prob}{A probability matrix generated by \code{build_prob_matrix}}

\item{tracing}{Print informational messages about how many people were newly
infected.}
}
\value{
A modified network graph after propagating the infections.
}
\description{
For each infected node on the network, find susceptible neighbors and
probabilistically spread the infection.
}
\examples{
# ADD_EXAMPLES_HERE
}
