% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect.R
\name{infect}
\alias{infect}
\title{Propagate infections on the network}
\usage{
infect(agents, neighbors, x0 = -2.556)
}
\arguments{
\item{agents}{A \code{data.table} containing agent information.}

\item{x0}{}

\item{prob}{A probability matrix generated by \code{build_prob_matrix}}
}
\value{
A modified network graph after propagating the infections.
}
\description{
For each infected node on the network, find susceptible neighbors and
probabilistically spread the infection.
}
\examples{
# ADD_EXAMPLES_HERE
}
