% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run a model}
\usage{
run_model(model, ticks, progress = FALSE, check_term = TRUE)
}
\arguments{
\item{model}{A model object returned from \code{\link[=setup_model]{setup_model()}}.}

\item{ticks}{Number of days to simulate.}

\item{progress}{Print progress messages for every tick
If it's \code{TRUE} or 1, a message is printed. If it's greater than 1
additional messages from \code{\link[=tick]{tick()}} are printed.}

\item{check_term}{Check populations in compartments and terminate early if
there are no agents in either E or I (so no further infections can happen).}
}
\value{
A named list with the resulting model after the ticks, and
a tibble with the counts of agents in "S", "E", "I", and "R" status.
}
\description{
Run a model for a specified number of ticks, where each tick is a day.
}
\examples{
# ADD_EXAMPLES_HERE
}
