% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agents.R
\name{setup_model}
\alias{setup_model}
\title{Set up a model}
\usage{
setup_model(
  agent_count,
  age_dist,
  p_female,
  seir_dist,
  p_med_cond,
  p_symptomatic,
  neighbors,
  p_rewire,
  probs = NULL,
  trans = NULL
)
}
\arguments{
\item{agent_count}{The number of agents to generate parameters for.}

\item{age_dist}{A function to sample from a probability distribution or
else a vector or list of ages. The vector will be repeated if
necessary to produce enough agents.}

\item{p_female}{A probability of being female or else a vector or list of
sexes ("M" or "F") or logical (\code{TRUE} for female, \code{FALSE} for male).
The vector will be repeated if necessary to produce enough agents.}

\item{seir_dist}{A function to sample from a probability distribution or
else a vector or list of SEIR statuses ("S", "E", "I", or "R"). The vector
will be repeated if necessary to produce enough agents.}

\item{p_med_cond}{A probability of having a medical condition or else a
vector or list of medical conditions (logical: \code{TRUE} or \code{FALSE}).
The vector will be repeated if necessary to produce enough agents.}

\item{p_symptomatic}{A probability of being symptomatic if infectious, or
else a vector or list of symptomatic status if infectious
(logical: \code{TRUE} or \code{FALSE}). The vector will be repeated if necessary to
produce enough agents.}

\item{neighbors}{The number of neighbor edges in the ring network graph to
initialize the small world network (2 = connect to both nearest neighbors
on the ring;  4 = connect to both neighbors and their neighbors, etc.).}

\item{p_rewire}{The probability to rewire an edge from a neighbor to a
random node..}

\item{probs}{A contagion probability matrix}

\item{trans}{Disease progression transition parameters}
}
\value{
A named list of the model: agents, network, contagion
probabilities for disease transmission, and transition probabilities for
progress to the next stage of the disease.
}
\description{
Initialize a model (agents and network).
}
\examples{
# ADD_EXAMPLES_HERE
}
